only use AstPythonVersion internally, move pep440 to external side
