+++ crates/ruff_workspace/src/pyproject.rs
@@ -5,7 +5,7 @@ use std::path::{Path, PathBuf};
 use anyhow::{Context, Result};
 use log::debug;
 use pep440_rs::VersionSpecifiers;
-use ruff_python_ast::python_version::PythonVersion;
+use ruff_linter::settings::types::PythonVersion;
 use serde::{Deserialize, Serialize};
 
 use crate::options::Options;
